Animator, Textures, Music
Animatiors and upadated textures all in. All the Music as well.
I imported Mike's new Animations for Jeremiah and tried to apply them,
but in doing so broke Jeremiah. I think it's just a tagging in the
code's inspector kind of problem. (He is a new model with a new
skelliton so had to replace his old prefab with this new one, can't get
him to spawn or control.)