Attempt #4: New File, remade, UI Menu
Redownloaded the entire project, because of an error. Should work now.
All that was added was folder-> Main Menu, into folder->UIImports. Only
possible error (to my knowledge) will be the fact that my in control
native and input interface files act though they are non-existent, but
I'm commiting this without those changes